ci: workflow updates + Strict-mode source fixes (continuation of 7287d8a)
The prior commit only staged the new/deleted files (a bad pathspec in
the git add call aborted staging of the rest). This commit carries the
actual workflow edits and the static_cast fixes described there:
avx512-compilation.yml, avx512-testing.yml, ci.yml, docs.yml, and the
gaussian_analysis.cpp/von_mises.cpp/tests/examples/tools cast fixes
that make the new Strict build type actually pass -Werror.

+++ examples/parallel_execution_demo.cpp
@@ -234,15 +234,14 @@ void demonstrate_cache_optimization() {
 void demonstrate_parallel_batch_fit() {
     print_separator("Parallel Batch Fitting: parallelBatchFit()");
 
-    std::cout
-        << "\nparallelBatchFit() fits independent datasets concurrently using the\n"
-        << "platform's thread pool. It is most useful when you have many datasets\n"
-        << "(e.g. simulation runs, cross-validation folds, streaming windows) that\n"
-        << "need to be fitted independently.\n\n";
+    std::cout << "\nparallelBatchFit() fits independent datasets concurrently using the\n"
+              << "platform's thread pool. It is most useful when you have many datasets\n"
+              << "(e.g. simulation runs, cross-validation folds, streaming windows) that\n"
+              << "need to be fitted independently.\n\n";
 
     std::mt19937 rng(42);
     constexpr size_t N_DATASETS = 50;
-    constexpr size_t N_OBS     = 500;
+    constexpr size_t N_OBS = 500;
 
     // --- Gaussian parallelBatchFit ------------------------------------------
     {
@@ -250,8 +249,8 @@ void demonstrate_parallel_batch_fit() {
         std::vector<std::vector<double>> datasets(N_DATASETS);
         std::uniform_real_distribution<double> mu_gen(-5.0, 5.0);
         std::uniform_real_distribution<double> sigma_gen(0.5, 3.0);
-        std::vector<stats::GaussianDistribution> true_params(N_DATASETS,
-            stats::GaussianDistribution::create(0.0, 1.0).unwrap());
+        std::vector<stats::GaussianDistribution> true_params(
+            N_DATASETS, stats::GaussianDistribution::create(0.0, 1.0).unwrap());
 
         for (size_t i = 0; i < N_DATASETS; ++i) {
             auto src = stats::GaussianDistribution::create(mu_gen(rng), sigma_gen(rng)).unwrap();
@@ -260,34 +259,41 @@ void demonstrate_parallel_batch_fit() {
         }
 
         // Sequential fit for timing comparison
-        std::vector<stats::GaussianDistribution> seq_results(N_DATASETS,
-            stats::GaussianDistribution::create(0.0, 1.0).unwrap());
+        std::vector<stats::GaussianDistribution> seq_results(
+            N_DATASETS, stats::GaussianDistribution::create(0.0, 1.0).unwrap());
         auto t0 = std::chrono::high_resolution_clock::now();
         for (size_t i = 0; i < N_DATASETS; ++i)
             seq_results[i].fit(datasets[i]);
         std::int64_t t_seq = std::chrono::duration_cast<std::chrono::microseconds>(
-                         std::chrono::high_resolution_clock::now() - t0).count();
+                                 std::chrono::high_resolution_clock::now() - t0)
+                                 .count();
 
         // Parallel fit
-        std::vector<stats::GaussianDistribution> par_results(N_DATASETS,
-            stats::GaussianDistribution::create(0.0, 1.0).unwrap());
+        std::vector<stats::GaussianDistribution> par_results(
+            N_DATASETS, stats::GaussianDistribution::create(0.0, 1.0).unwrap());
         t0 = std::chrono::high_resolution_clock::now();
         stats::GaussianDistribution::parallelBatchFit(datasets, par_results);
         std::int64_t t_par = std::chrono::duration_cast<std::chrono::microseconds>(
-                         std::chrono::high_resolution_clock::now() - t0).count();
+                                 std::chrono::high_resolution_clock::now() - t0)
+                                 .count();
 
         // Verify: parallel and sequential results should match
         bool match = true;
         for (size_t i = 0; i < N_DATASETS; ++i) {
-            if (std::abs(par_results[i].getMean() - seq_results[i].getMean()) > 1e-10)
-                { match = false; break; }
+            if (std::abs(par_results[i].getMean() - seq_results[i].getMean()) > 1e-10) {
+                match = false;
+                break;
+            }
         }
 
         std::cout << std::fixed << std::setprecision(1);
         std::cout << "Gaussian: " << N_DATASETS << " datasets x " << N_OBS << " obs each\n";
         std::cout << "  Sequential fit: " << t_seq << " us\n";
         std::cout << "  Parallel fit:   " << t_par << " us"
-                  << "  (" << static_cast<double>(t_seq) / std::max<std::int64_t>(t_par, 1) << "x)\n";
+                  << "  ("
+                  << static_cast<double>(t_seq) /
+                         static_cast<double>(std::max<std::int64_t>(t_par, 1))
+                  << "x)\n";
         std::cout << "  Results match sequential: " << (match ? "yes" : "NO") << "\n";
     }
 
@@ -300,12 +306,13 @@ void demonstrate_parallel_batch_fit() {
             datasets[i] = src.sample(rng, N_OBS);
         }
 
-        std::vector<stats::ExponentialDistribution> par_results(N_DATASETS,
-            stats::ExponentialDistribution::create(1.0).unwrap());
+        std::vector<stats::ExponentialDistribution> par_results(
+            N_DATASETS, stats::ExponentialDistribution::create(1.0).unwrap());
         auto t0 = std::chrono::high_resolution_clock::now();
         stats::ExponentialDistribution::parallelBatchFit(datasets, par_results);
         std::int64_t t_par = std::chrono::duration_cast<std::chrono::microseconds>(
-                         std::chrono::high_resolution_clock::now() - t0).count();
+                                 std::chrono::high_resolution_clock::now() - t0)
+                                 .count();
 
         std::cout << "Exponential: " << N_DATASETS << " datasets x " << N_OBS << " obs each\n";
         std::cout << "  Parallel fit: " << t_par << " us\n";
